Added the perl part of the solution "as is" that I submitted to Woven for the question about removing files older than some number of days and matching a pattern.
Here was the question:
Cleaning up old files
We've run into a problem – client file uploads aren't being
automatically removed correctly, so we are running out of disk space on
one of the servers. We could remove these one-by-one, but this problem
would come back in the future and there are a lot of files to go
through.

Any file with a date over 3 months old should be deleted, but there are
files dating back almost a year. And the naming convention changed
sometime around October or November 2018, so you'll have to account for
2 different name formats. Note that other processes could touch the
files, so you can't rely on the last modified time.

Client upload filenames have a date in ISO 8601 format. Here are some
example filenames and the expected result after running the script,
assuming today’s date is 2019-03-01:

uploads/aperture-2018-04-01-spend_data.csv  # should be deleted
uploads/cyberdyne/2019-02-01-claims.csv     # should not be deleted
uploads/cyberdyne/2018-09-01-claims.csv     # should be deleted
You can download a directory with many example files to test your
script.

Time guidance
Aim to take between 25-45 minutes for this challenge. Set a timer now
for 45 minutes to gauge the time you have spent on this challenge.

If you have spent 45 or more minutes:

Stop working
Document where you are at in the "Your Notes" section in the upper right
corner
Go on to the next challenge. It is independent from any solution you got
on this challenge.
You can return to any incomplete challenges if you have time left over
at the end.